+++ src/routes/Setting/Board/index.tsx
@@ -1,32 +1,56 @@
-import React, { useState } from "react";
-import styled from "styled-components";
+import React, { useCallback, useState } from "react";
 import BoardCategoryList from "../../../components/board/BoardCategoryList";
 import { PlusIcon } from "../../../components/Icons/SettingIcons";
 import { useBoardCategoryCreate } from "../../../hooks/board/category/useBoardCategoryCreate";
 import { useBoardCategoryList } from "../../../hooks/board/category/useBoardCategoryList";
-import COLOR from "./../../../constants/color";
-import PageTitle from "./../../../components/common/pagetitle/index";
 import FONT from "./../../../constants/fonts";
+import Modal from "../../../components/common/modal";
+import {
+  settingCategoryId,
+  settingCategoryName,
+  settingFromState,
+} from "./../../../states/board";
+import {
+  CategoryListBox,
+  SettingButton,
+  SettingContainer,
+  SettingInput,
+  SettingTitle,
+  SettingTop,
+} from "./style";
+import { useRecoilState, useRecoilValue } from "recoil";
+import { useBoardCategoryUpdate } from "../../../hooks/board/category/useBoardCategoryUpdate";
 
 const BoardManagement = () => {
   const [boardName, setBoardName] = useState<string>("");
   const createMutation = useBoardCategoryCreate(boardName);
 
+  const { boardCategoryList } = useBoardCategoryList();
+
+  const [isOpenModal, setIsOpenModal] = useState(false);
+  const [settingForm, setSettingForm] = useRecoilState(settingFromState);
+
+  const categoryId = useRecoilValue(settingCategoryId);
+  const [categoryName, setCategoryName] = useRecoilState(settingCategoryName);
+
+  const categoryUpdate = useBoardCategoryUpdate(categoryId, categoryName);
+
   const handleCreateCategory = () => {
     createMutation.mutate();
     setBoardName("");
-    setSettingForm(false);
+    setIsOpenModal(false);
   };
 
-  const { boardCategoryList } = useBoardCategoryList();
-  const [settingForm, setSettingForm] = useState(false);
+  const handleClickCreateModal = useCallback(() => {
+    setIsOpenModal(!isOpenModal);
+  }, [isOpenModal]);
 
   return (
     <>
       <SettingContainer style={{ width: "30%" }}>
         <SettingTop>
           <SettingTitle style={FONT.HEADING}>게시판</SettingTitle>
-          <button onClick={() => setSettingForm((prev) => !prev)}>
+          <button onClick={handleClickCreateModal}>
             <PlusIcon />
           </button>
         </SettingTop>
@@ -37,17 +61,9 @@ const BoardManagement = () => {
         </CategoryListBox>
       </SettingContainer>
 
-      {settingForm && (
-        <SettingContainer style={{ width: "50%" }}>
-          <SettingTop>
-            <SettingTitle style={FONT.HEADING}>설정</SettingTitle>
-            <SettingButton
-              style={FONT.SUBTITLE2}
-              onClick={handleCreateCategory}
-            >
-              저장
-            </SettingButton>
-          </SettingTop>
+      {isOpenModal && (
+        <Modal onClickToggleModal={handleClickCreateModal}>
+          <SettingTitle>게시판 생성</SettingTitle>
           <SettingInput
             type="text"
             value={boardName}
@@ -56,43 +72,38 @@ const BoardManagement = () => {
               setBoardName(e.currentTarget.value)
             }
           />
+          <SettingButton style={FONT.SUBTITLE2} onClick={handleCreateCategory}>
+            저장
+          </SettingButton>
+        </Modal>
+      )}
+
+      {settingForm && (
+        <SettingContainer style={{ width: "50%" }}>
+          <SettingTop>
+            <SettingTitle style={FONT.HEADING}>설정</SettingTitle>
+            <SettingButton
+              style={FONT.SUBTITLE2}
+              onClick={() => {
+                categoryUpdate.mutate();
+                setSettingForm((prev) => !prev);
+              }}
+            >
+              변경사항 저장
+            </SettingButton>
+          </SettingTop>
+          <SettingInput
+            type="text"
+            value={categoryName}
+            style={FONT.SUBTITLE2}
+            onChange={(e: React.FormEvent<HTMLInputElement>) =>
+              setCategoryName(e.currentTarget.value)
+            }
+          />
         </SettingContainer>
       )}
     </>
   );
 };
-
-const SettingContainer = styled.div`
-  height: 100%;
-  background: ${COLOR.WHITE};
-  padding: 2rem;
-  border-radius: 1rem;
-`;
-
-const SettingTop = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 2rem;
-`;
-
-const SettingTitle = styled.div`
-  color: ${COLOR.BLACK};
-`;
-
-const SettingInput = styled.input`
-  width: 100%;
-  height: 3rem;
-  border-bottom: 1px solid ${COLOR.GRAY0};
-`;
-
-const SettingButton = styled.button`
-  color: ${COLOR.GREEN4};
-`;
-
-const CategoryListBox = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
 
 export default BoardManagement;
